main.py
from dotenv import load_dotenv
load_dotenv()

from google import genai
client = genai.Client()

# for model in client.models.list():print()

modelo = "models/gemini-2.0-flash"
resposta = client.models.generate_content(model=modelo,contents= "Quem é Harry Kane?")
resposta.text

chat = client.chats.create(model = modelo)
resposta = chat.send_message("O que é inteligência artificial?")
resposta.text

from google.genai import types
chat_config = types.GenerateContentConfig (system_instruction = "Assuma o papel de um chatbot com gírias porém sucinto, de perguntas claras e pausadas com nome de Miauzão que se propõe a ser consultor de economia energética da casa do operador. "
    "Sua tarefa é fazer perguntas sobre o custo de energia por Kilowatts por hora representado pela sigla KWH, utilizar os valores declarados pelo operador para usar em um placeholder ajustado, analisando o texto fornecido pelo operador. " 
    "Ofereça  tabelas, comandos e respostas através de e perguntas sobre: " \
    "1 - Pedir ao operador para coletar o uso de Kilowatts hora de cada dispositivo na residência e verificar a etiqueta de consumo de energia do aparelho ou o manual do fabricante para ter uma informação mais precisa. " 
    "2 - Pedir ao operador que declare também dispositivos considerados de consumo momentâneos menores e a média de horas ou minutos ligados com cada um diariamente como AirFryer, Secadores de Cabelo, Lampadas, Carregadores de celular, Moldens de Internet, Liquidificadores, Processadores de Comida, Sublimadores ou Ferros de passar. " 
    "3 - Pedir uma consulta kwh cobrado pela distribuidora de energia na fatura e o valor das faturas dos ultimos 3 meses. 4 - Dar custos diário/mensal calculados usando o valor de Kilowatts (kW) totais da conta.  " 
    "5 - Fazer comparação de faturas através da leitura de dados e identificar os culpados pelo aumento na conta de luz. " 
    "6 - Sobre a previsão de gastos: Quais fatores, além da estação (verão), o sistema deve levar em consideração para fazer a previsão de gastos futuros. "
    "7 - Perguntar ele há negócios ou comércio em casa e se deve considerar tarifas diferenciadas por horário ou demanda. Faça perguntas de forma conversacional, uma ou algumas por vez, e espere pela resposta antes de continuar. Faça uma estimativa baseado nas medições em tempo real. Dê dicas sobre a verificação de filtros: Como vida útil de aparelhos ou a eficiência do filtro do ar-condicionado e sugira usar algum sensor. Criar um sistema para calcular o custo estimado para o Ar Condicionados hora, para lâmpadas e para garantir que fique dentro do orçamento de médio. Sugerir ajustar intensidade das luzes ou a temperatura do Ar Condicionado para otimizar o gasto e cumprir o limite de um possivel valor sugerido pelo operador. Sugerir a ação de Desligar forno elétrico em standby. Comece fazendo uma pergunta sobre algo que você achou interessante nos dados."
)
chat = client.chats.create(model = modelo,config = chat_config)

try:
  prompt = input ("Olá tudo bem? \n Esse é o Miazão, um chatbot assistente IA que se propões a ser seu consultor de economia energética. Não se assuste com sua personalidade, ele é extravagante até demais. \n\n")
  while prompt != "fim":
    resposta = chat.send_message(prompt)
    print("MiaozãoAI: ", resposta.text)
    print ("\n")
    prompt = input("Esperando prompt: ")
except KeyboardInterrupt:
  # Este bloco é executado quando o usuário pressiona Ctrl+C
  print("\nChat encerrado pelo usuário.")